Добавлен viewHolder, скругление с помощью Glide
