+++ app/src/main/res/layout/track_view.xml
@@ -1,27 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools"
     android:orientation="horizontal"
     android:paddingHorizontal="@dimen/padding_editText_horizontal"
     android:paddingVertical="@dimen/padding_editText_vertical">
 
-    <androidx.cardview.widget.CardView
+
+    <ImageView
+        android:id="@+id/track_image"
         android:layout_width="@dimen/size_track_image"
         android:layout_height="@dimen/size_track_image"
-        app:cardBackgroundColor="@android:color/transparent"
-        app:cardCornerRadius="2dp"
-        app:cardElevation="0dp">
-
-        <ImageView
-            android:id="@+id/track_image"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:scaleType="centerCrop"
-            android:src="@drawable/track_placeholder" />
-    </androidx.cardview.widget.CardView>
+        android:scaleType="centerCrop"
+        android:src="@drawable/track_placeholder" />
 
     <LinearLayout
         android:layout_width="0dp"
@@ -33,13 +26,13 @@
 
         <TextView
             android:id="@+id/track_name"
+            style="@style/TextView_track"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            style="@style/TextView_track"
-            tools:text="Smells Like Teen Spirit"
             android:textAppearance="?android:attr/textAppearanceMedium"
             android:textColor="?attr/colorOnSecondary"
-            android:textSize="@dimen/secondary_text_size" />
+            android:textSize="@dimen/secondary_text_size"
+            tools:text="Smells Like Teen Spirit" />
 
         <LinearLayout
             android:layout_width="wrap_content"
@@ -48,26 +41,26 @@
 
             <TextView
                 android:id="@+id/artist_name"
+                style="@style/TextView_track"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
-                style="@style/TextView_track"
                 android:layout_marginEnd="4dp"
                 android:layout_weight="1"
-                tools:text="Nirvana"
                 android:drawableEnd="@drawable/dot"
                 android:drawablePadding="4dp"
                 android:textColor="?attr/colorIconSecondary"
-                android:textSize="@dimen/third_text_size" />
+                android:textSize="@dimen/third_text_size"
+                tools:text="Nirvana" />
 
             <TextView
                 android:id="@+id/track_time"
+                style="@style/TextView_track"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                style="@style/TextView_track"
-                tools:text="5:01"
                 android:textAppearance="?android:attr/textAppearanceSmall"
                 android:textColor="?attr/colorIconSecondary"
-                android:textSize="@dimen/third_text_size" />
+                android:textSize="@dimen/third_text_size"
+                tools:text="5:01" />
         </LinearLayout>
     </LinearLayout>
 
@@ -77,5 +70,5 @@
         android:layout_height="wrap_content"
         android:layout_gravity="center"
         android:paddingHorizontal="@dimen/padding_icon"
-        android:src="@drawable/right_arrow"/>
+        android:src="@drawable/right_arrow" />
 </LinearLayout>
